% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{anorexia_weight_change_df}
\alias{anorexia_weight_change_df}
\title{Anorexia Weight Change}
\format{
A data frame with 72 observations and 3 variables:
\describe{
  \item{Treat}{Factor indicating the treatment type (3 levels)}
  \item{Prewt}{Numeric vector indicating the patient's weight before treatment (in kilograms)}
  \item{Postwt}{Numeric vector indicating the patient's weight after treatment (in kilograms)}
}
}
\source{
Data taken from the MASS package version 7.3-65.
}
\usage{
data(anorexia_weight_change_df)
}
\description{
This dataset, anorexia_weight_change_df, is a data frame containing weight change data
for young female anorexia patients. It includes pre- and post-treatment weights,
along with the type of treatment administered.
}
\details{
The dataset name has been kept as 'anorexia_weight_change_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
DigestiveDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
