% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{campylobacter_infections_ts}
\alias{campylobacter_infections_ts}
\title{Campylobacter Infections Time Series}
\format{
A time series object (`ts`) with 140 observations:
\describe{
  \item{Start}{c(1990, 1)}
  \item{End}{c(2000, 10)}
  \item{Frequency}{13 (observations per year)}
}
}
\source{
Data taken from the \pkg{tscount} package version 1.4.3. Original source: Ferland, R., Latour, A. and Oraichi, D.,
"Integer-valued GARCH process". \emph{Journal of Time Series Analysis}, 2006; 27(6): 923--942.
}
\usage{
data(campylobacter_infections_ts)
}
\description{
This dataset, campylobacter_infections_ts, is a time series object containing the number of cases
of campylobacter infections in northern Quebec (Canada), recorded in four-week intervals
from January 1990 to October 2000. Campylobacterosis is an acute bacterial infectious disease
attacking the digestive system.
}
\details{
The dataset name has been kept as 'campylobacter_infections_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
DigestiveDataSets package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
