% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestivedatasets-package.R
\docType{package}
\name{DigestiveDataSets}
\alias{DigestiveDataSets}
\alias{DigestiveDataSets-package}
\title{DigestiveDataSets: A Curated Collection of Digestive System and Gastrointestinal Disease Datasets}
\description{
{
This package provides a wide variety of datasets focused on the digestive system, stomach, intestines, liver, pancreas, and associated diseases, including clinical trials, observational studies, experimental datasets, cohort data, and case series involving gastrointestinal disorders such as gastritis, ulcers, pancreatitis, liver cirrhosis, colon cancer, colorectal conditions, Helicobacter pylori infection, irritable bowel syndrome, intestinal infections, and post-surgical outcomes.
}
}
\details{
DigestiveDataSets: A Curated Collection of Digestive System and Gastrointestinal Disease Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Curated Collection of Digestive System and Gastrointestinal Disease Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/digestivedatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
