% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ecoli_infections_df}
\alias{ecoli_infections_df}
\title{E. coli Infections Time Series}
\format{
A data frame with 646 observations and 3 variables:
\describe{
  \item{year}{Numeric value indicating the year of observation}
  \item{week}{Numeric value indicating the week of observation}
  \item{cases}{Numeric value indicating the number of reported E. coli cases}
}
}
\source{
Data taken from the \pkg{tscount} package version 1.4.3.
}
\usage{
data(ecoli_infections_df)
}
\description{
This dataset, ecoli_infections_df, is a data frame containing the weekly number of reported disease cases
caused by Escherichia coli in the state of North Rhine-Westphalia (Germany) from January 2001 to May 2013,
excluding cases of EHEC and HUS.
}
\details{
The dataset name has been kept as 'ecoli_infections_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
DigestiveDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
