% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{gi_damage_prevention_df}
\alias{gi_damage_prevention_df}
\title{Gastrointestinal Damage Prevention}
\format{
A data frame with 198 observations and 3 variables:
\describe{
  \item{study}{Factor indicating the clinical trial (4 levels)}
  \item{treatment}{Factor indicating the treatment group (2 levels: control or Misoprostol)}
  \item{classification}{Ordered factor indicating the degree of gastrointestinal damage (5 levels)}
}
}
\source{
Data taken from the HSAUR3 package version 1.0-15.
}
\usage{
data(gi_damage_prevention_df)
}
\description{
This dataset, gi_damage_prevention_df, is a data frame containing results from four
randomised clinical trials on the prevention of gastrointestinal damages by Misoprostol,
reported by Lanza et al. (1987–1989).
}
\details{
The dataset name has been kept as 'gi_damage_prevention_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
DigestiveDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
