% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{helicobacter_children_tbl_df}
\alias{helicobacter_children_tbl_df}
\title{Helicobacter pylori Infection in Preschoolers}
\format{
A tibble with 863 observations and 2 variables:
\describe{
  \item{ulcer}{Factor with 2 levels indicating parental history of duodenal or gastric ulcer}
  \item{infected}{Factor with 2 levels indicating Helicobacter pylori infection status}
}
}
\source{
Data taken from the package pubh version 2.0.0.
}
\usage{
data(helicobacter_children_tbl_df)
}
\description{
This dataset, helicobacter_children_tbl_df, is a tibble containing the prevalence of
Helicobacter pylori infection in preschool children according to parental history
of duodenal or gastric ulcer.
}
\details{
The dataset name has been kept as 'helicobacter_children_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
DigestiveDataSets package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
