% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{intestinal_surgery_df}
\alias{intestinal_surgery_df}
\title{Satellite Tumors in GI Surgery}
\format{
A data frame with 844 observations and 2 variables:
\describe{
  \item{n}{Numeric value representing the number of satellites removed}
  \item{s}{Numeric value representing the number of malignant satellites found}
}
}
\source{
Data taken from the \pkg{deconvolveR} package version 1.2-1. Original source:
Efron, B. (2016). "Empirical Bayes deconvolution estimates". \emph{Biometrika}, 103(1), 1--20.
}
\usage{
data(intestinal_surgery_df)
}
\description{
This dataset, intestinal_surgery_df, is a data frame containing
intestinal surgery data from 844 cancer patients. The data consists of pairs
(n_i, s_i) where n_i is the number of satellites removed and s_i is the number
of satellites found to be malignant.
}
\details{
The dataset name has been kept as 'intestinal_surgery_df' to avoid
confusion with other datasets in the R ecosystem. This naming convention helps
distinguish this dataset as part of the DigestiveDataSets package and assists
users in identifying its specific characteristics. The suffix 'df' indicates
that the dataset is a data frame. The original content has not been modified
in any way.
}
