% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{weight_loss_df}
\alias{weight_loss_df}
\title{Obese Patient Weight Loss Data}
\format{
A data frame with 52 observations and 2 variables:
\describe{
  \item{Days}{Integer vector indicating the number of days since the beginning of the programme}
  \item{Weight}{Numeric vector indicating the weight (in kilograms) of the patient at each time point}
}
}
\source{
Data taken from the MASS package version 7.3-65.
}
\usage{
data(weight_loss_df)
}
\description{
This dataset, weight_loss_df, is a data frame containing the weight, in kilograms,
of an obese patient measured at 52 time points over an 8-month period as part of a
weight rehabilitation programme.
}
\details{
The dataset name has been kept as 'weight_loss_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
DigestiveDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
